requirements.txt
fastapi
uvicorn
streamlit
requests
scikit-learn
python-dotenv
groq
numpy
matplotlib